agendas-api/requirements.txt
# requirements.txt (agendas-api)

Django==4.2.5
djangorestframework==3.14.0
drf-yasg==1.21.7
django-cors-headers==4.2.0
gunicorn==21.2.0
Faker==37.0.0
python-dateutil
typing-extensions